requirements.txt
fastapi
uvicorn[standard]
google-generativeai
Pillow
Jinja2